Work on input & clientside console features detection

+++ ConsoleSurf.Server/tiocsti.c
@@ -7,6 +7,6 @@
 // $ gcc -shared -o tiocsti.so tiocsti.o
 #include <sys/ioctl.h>
 
-extern int get_tiocsti() {
+extern unsigned long get_tiocsti() {
     return TIOCSTI;
 }
